api/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
marshmallow-sqlalchemy==0.23.0
